Cleanup AppArmor profiles as per intrigeri's pull request to upstream.
commit 6552a2ea511cdc268f020984e1ce745d91562f94
Author: intrigeri <intrigeri@boum.org>
Date:   Wed Jul 23 18:48:40 2014 +0000

Submitted as https://github.com/micahflee/torbrowser-launcher/pull/111

These updated profiles clean up and refactor many parts of the profiles,
fix a bunch of bugs, and make it fit for non-Ubuntu distros. Tested on
Debian Wheezy only, as torbrowser-launcher is broken with the version of
Twisted that's in Debian unstable currently.