AppArmor: allow Firefox to get entropy from @{PROC}/sys/kernel/random/uuid.
First, note that this file returns a different uuid each time it's read.

I'm not sure what Firefox does when it can't get a random UUID from there. If it
falls back to a fixed value, this might have problematic security implications:
randomness is sometimes useful for security purposes. The only place I could
find this file being used in the Firefox source tree is actually in the
`arc4_seed_proc_sys_kernel_random_uuid` function. If that's really why Firefox
tries to access this file, then I don't think we should block it.

I've not seen that file being accessed by Firefox personally. But given
troubadoour added it, and it can be found in the Firefox source code, with the
above reasoning in mind, it seems that the safest thing to do is to allow
Firefox to get the randomness it needs.