AppArmor: allow Firefox to read processes and tasks' stats.
This partially reverts commit 04b24660, that made the opposite change for
reasons that are unknown to me.

stat files are used in the JiffiesSinceBoot function
(xpcom/ds/TimeStamp_posix.cpp), which is used to compute process lifetime.
The consequences of blocking this access are unclear to me: it might plug issues
wrt. anonymity that the Tor Browser team would have missed (ask them?), but it
can as well introduce security issues by forcing Firefox to downgrade to worse
sources of information. If crypto is in play there, we would be playing
a dangerous game by blocking Firefox from accessing this information.