Split print calls on newlines
If torbrowser-launcher cannot write to stdout, e.g. because it was
started in the background and the controlling terminal has been closed
or because it was started from a desktop environment launcher whose
stdout has been closed, it crashes after updating the GnuPG key.

This is due to print() crashing the program if stdout isn't writeable
and the string to print contains a newline.

To work around the problem, split the strings containing newlines into
several calls to print().

See also the upstream bug at https://bugs.python.org/issue32345

Closes #298