AppArmor: adjust Firefox binary path for Tor Browser 8.0a10.
At this point it seems unlikely that the develop branch will be released
before Tor Browser 8.0 so here we go, let's get ready.

Note that I could have written firefox{,.real} instead, to support both Tor
Browser 7.5 and 8.0, but then we would have to open the profile more broadly so
the new shell wrapper installed as "firefox" by Tor Browser 8.0a10 can do its
job. This does not seem worth the hassle and will be fine as long as this new
torbrowser-launcher is released approximately at the same time as, or after, Tor
Browser 8.