Open additional URLs with the same TorBrowser instance
When we looked at the history of issues about opening links with TorBrowser, we
found #103 where initially the feature to open links from other applications
was added. Then some months later, the feature was removed again, as it was not
working because of #157 and #175. The issues was back then (~4years ago), that
when users had a normal Firefox running, urls got opened in the normal firefox
insead of TorBrowser. That was because TorBrowser had the --no-remote flag set
deep down in their code.
In the meanwhile TorBrowser removed the explicit --no-remote flag, so we are
able to use --allow-remote again.
On top of that there is another issue when we want open urls with TorBrowser:
If the LOGNAME is the same as the LOGNAME of the opened firefox, urls are still
opened in the open firefox. But if we set the LOGNAME envrionment variable
explicitly, we can make sure, that the url opens in the TorBrowser instead of
a running Firefox.

Should fix: #245, #259, #380