Add X abstractions, override X abstraction
X abstractions are needed since we use X Server for the display bits for Firefox Browser.

On Ubuntu/Lubuntu though, in 21.04 and up, the "write" permission is yanked from the X display sockets in `/tmp/.X11-unix/*` though, and is needed, so we'll override it for now in-profile.  This addresses https://github.com/micahflee/torbrowser-launcher/issues/588.