AppArmor: remove duplicate rule
We include the gnome abstraction, which itself includes the X abstraction, so
at least in theory including the X abstraction explicitly should be a no-op.

It looks like adding this line was done based on an incorrect assumption (that
Lubuntu does not include the gnome abstraction), and even if that assumption
were correct, adding this line would not fix the problem (because the parser
would fail to compile the profile regardless).

 - https://gitlab.torproject.org/tpo/applications/torbrowser-launcher/-/merge_requests/16#note_3001282
 - https://gitlab.torproject.org/tpo/applications/torbrowser-launcher/-/merge_requests/16#note_3001283

If there's ever evidence of a system where this line actually solves a problem,
we can reconsider, and I'll be very curious to understand why.