Don't grant full D-Bus session bus access
This very broad access was added in e049fdcc769e9b71078d4bc0256441286a53d24c
without a clear description of what problem it fixed. It's very broad and
dangerous, as in: it probably allows an adversary who took control of the
browser to escape the sandbox is a variety of ways.

Instead, let's only allow access to the bus itself. Once we have bug reports
that help us understand which additional permissions may be needed, we
can reconsider. E.g. we can add fine-grained D-Bus mediation on systems
that support it, or simply revert this commit if it turns out it's the
best solution to a well-defined problem.

References:

- https://gitlab.torproject.org/tpo/applications/torbrowser-launcher/-/merge_requests/16#note_3001291
- https://gitlab.torproject.org/tpo/applications/torbrowser-launcher/-/issues/23#note_3093839

Closes tpo/applications/torbrowser-launcher#23