AppArmor: allow unprivileged user namespaces
Firefox uses userns to set up its own sandboxing.

On Debian, AppArmor was already allowing this by default, until a recent
upload (that is now in Trixie) updated the features pinning to a version that
now mediates usage of userns, so this functionality is now blocked by profiles
that don't explicitly allow it. Let's repair this.

Also reported as Debian#1098845.